# contracts/openapi/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LibraMAS Recommendation API",
    "version": "0.1.0-draft",
    "description": "Versioned contract baseline. Health and explicitly injected G3/G5 slices may be enabled gate by gate; the default runtime remains fail-closed and does not claim that recommendation HTTP is enabled."
  },
  "servers": [
    {"url": "http://localhost:8000", "description": "Local development"}
  ],
...
  "paths": {
...
    "/api/v1/recommendation-tasks": {
      "post": {
        "tags": ["Recommendation"],
        "summary": "Create or replay a recommendation task",
        "operationId": "recommendation_create_task_v1",
        "security": [{"bearerAuth": []}, {"demoUserAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/ClientRequestId"},
          {"$ref": "#/components/parameters/IdempotencyKey"},
          {"$ref": "#/components/parameters/DemoUserId"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RecommendationTaskCreateRequest"}}}
        },
        "responses": {
          "201": {
            "description": "New completed, degraded, or clarification-waiting task",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
              "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RecommendationExecutionResponse"}}}
          },
          "200": {
            "description": "Idempotently replayed task",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
              "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RecommendationExecutionResponse"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "422": {"$ref": "#/components/responses/UnprocessableEntity"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"},
          "504": {"$ref": "#/components/responses/GatewayTimeout"}
        }
      }
    },
    "/api/v1/recommendation-tasks/{task_id}/clarifications": {
      "post": {
        "tags": ["Recommendation"],
        "summary": "Resume the same task with clarification answers",
        "operationId": "recommendation_submit_clarification_v1",
        "security": [{"bearerAuth": []}, {"demoUserAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/ClientRequestId"},
          {"$ref": "#/components/parameters/TaskId"},
          {"$ref": "#/components/parameters/IdempotencyKey"},
          {"$ref": "#/components/parameters/DemoUserId"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ClarificationRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Resumed task state",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
              "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RecommendationExecutionResponse"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "422": {"$ref": "#/components/responses/UnprocessableEntity"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"},
          "504": {"$ref": "#/components/responses/GatewayTimeout"}
        }
      }
    },
    "/api/v1/recommendation-tasks/{task_id}": {
      "get": {
        "tags": ["Recommendation"],
        "summary": "Read recommendation task state",
        "operationId": "recommendation_get_task_v1",
        "security": [{"bearerAuth": []}, {"demoUserAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/ClientRequestId"},
          {"$ref": "#/components/parameters/TaskId"},
          {"$ref": "#/components/parameters/DemoUserId"}
        ],
        "responses": {
          "200": {
            "description": "Current task state",
            "headers": {
              "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
              "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RecommendationTaskStatusResponse"}}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "422": {"$ref": "#/components/responses/UnprocessableEntity"},
          "429": {"$ref": "#/components/responses/TooManyRequests"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"},
          "504": {"$ref": "#/components/responses/GatewayTimeout"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"},
      "demoUserAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Demo-User-Id",
        "description": "Local demo identity only. The server accepts it only when APP_ENV=demo and demo identity is explicitly enabled; it never grants research_admin privileges."
      }
    },
    "parameters": {
      "ClientRequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "schema": {"type": "string", "format": "uuid"},
        "description": "Optional client-generated request UUID. The server validates and echoes it; when absent, the server generates one."
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {"type": "string", "minLength": 8, "maxLength": 255},
        "description": "Stable key for safe replay of a write request."
      },
      "DemoUserId": {
        "name": "X-Demo-User-Id",
        "in": "header",
        "required": false,
        "schema": {"type": "integer", "minimum": 1},
        "description": "User-scoped routes only. Accepted only when APP_ENV=demo and demo identity is explicitly enabled; production derives identity from bearer authentication. This header never grants research_admin privileges."
      },
      "TaskId": {"name": "task_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
...
    },
    "headers": {
      "XRequestId": {
        "description": "The validated or server-generated request UUID. It is returned on every operation response.",
        "schema": {"type": "string", "format": "uuid"}
      },
      "XTraceId": {
        "description": "The trace UUID used to correlate the complete operation. It is returned on every operation response.",
        "schema": {"type": "string", "format": "uuid"}
      },
      "IdempotencyReplayed": {
        "description": "True when an idempotent write response was replayed from the original result; false otherwise.",
        "schema": {"type": "boolean"}
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed JSON or invalid HTTP protocol input",
        "headers": {
          "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
          "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
          "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
      },
      "Unauthorized": {
        "description": "Authentication is required or invalid",
        "headers": {
          "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
          "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
          "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
      },
      "Forbidden": {
        "description": "The caller cannot access the requested data or score view",
        "headers": {
          "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
          "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
          "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
      },
      "NotFound": {
        "description": "The requested resource was not found",
        "headers": {
          "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
          "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
          "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
      },
      "Conflict": {
        "description": "Context version or idempotency conflict",
        "headers": {
          "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
          "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
          "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
      },
      "UnprocessableEntity": {
        "description": "The request violates a domain constraint",
        "headers": {
          "X-Request-Id": {"$ref": "#/components/headers/XRequestId"},
          "X-Trace-Id": {"$ref": "#/components/headers/XTraceId"},
          "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
      },
...
    },
    "schemas": {
...
      "RecommendationTaskCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["request_id", "session_id", "scene"],
        "properties": {
          "request_id": {"type": "string", "format": "uuid"},
          "user_id": {"type": ["integer", "null"], "minimum": 1},
          "session_id": {"type": "string", "format": "uuid"},
          "scene": {"$ref": "#/components/schemas/TriggerScene"},
          "input_text": {"type": ["string", "null"], "maxLength": 2000},
          "requested_resource_types": {
            "type": "array",
            "uniqueItems": true,
            "items": {"$ref": "#/components/schemas/ResourceType"},
            "default": ["BOOK", "PAPER"]
          },
          "requested_output_type": {"oneOf": [{"$ref": "#/components/schemas/OutputType"}, {"type": "null"}]},
          "source_resource_id": {"type": ["integer", "null"], "minimum": 1},
          "source_item_id": {"type": ["integer", "null"], "minimum": 1},
          "as_of_time": {"type": ["string", "null"], "format": "date-time"},
          "constraints": {"type": "object", "additionalProperties": true},
          "limit": {"type": "integer", "minimum": 1, "maximum": 20, "default": 10}
        }
      },
...
      "RecommendationExecutionResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["task_id", "trace_id", "status", "context_version", "decision", "warnings"],
        "properties": {
          "task_id": {"type": "string", "format": "uuid"},
          "record_id": {"type": ["integer", "null"], "minimum": 1},
          "trace_id": {"type": "string", "format": "uuid"},
          "status": {"$ref": "#/components/schemas/TaskStatus"},
          "context_version": {"type": "integer", "minimum": 1},
          "evaluation_at": {"type": "string", "format": "date-time"},
          "decision": {"$ref": "#/components/schemas/InteractionDecision"},
          "groups": {"type": "array", "items": {"$ref": "#/components/schemas/RecommendationGroup"}},
          "items": {"type": "array", "items": {"$ref": "#/components/schemas/RecommendationItem"}},
          "questions": {"type": "array", "items": {"$ref": "#/components/schemas/ClarificationQuestion"}},
          "warnings": {"type": "array", "items": {"$ref": "#/components/schemas/WarningCode"}},
          "agent_actions": {"type": "array", "items": {"$ref": "#/components/schemas/AgentAction"}},
          "versions": {"$ref": "#/components/schemas/VersionBundle"}
        }
      },
...
      "ResourceType": {"type": "string", "enum": ["BOOK", "PAPER"]},
      "TriggerScene": {"type": "string", "enum": ["HOME", "SEARCH_AFTER", "RESOURCE_DETAIL", "FEEDBACK_REFRESH", "EXPLANATION"]},
...
    }
  },
  "security": [{"bearerAuth": []}]
}